requirements.txt
djangorestframework
djangorestframework-simplejwt
drf-nested-routers
